Trait libimagerror::iter::TraceIterator

pub trait TraceIterator<T, E>: Iterator<Item=Result<T, E>> + Sized {
    fn trace_unwrap(self) -> UnwrapWith<Self, fn(E)> where E: Error { ... }
    fn on_err<F>(self, f: F) -> OnErr<Self, F> where F: FnMut(&E) { ... }
    fn unwrap_with<F>(self, f: F) -> UnwrapWith<Self, F> where F: FnMut(E) { ... }
}

This trait provides methods that make it easier to work with iterators that yield a Result.

Provided Methods

fn trace_unwrap(self) -> UnwrapWith<Self, fn(E)> where E: Error

Creates an iterator that yields the item in each Ok item, while filtering out the Err items. Each filtered Err will be trace-logged with [::trace::trace_error].

As with all iterators, the processing is lazy. If you do not use the result of this method, nothing will be passed to ::trace::trace_error, no matter how many Err items might be present.

fn on_err<F>(self, f: F) -> OnErr<Self, F> where F: FnMut(&E)

Takes a closure and creates an iterator that will call that closure for each Err element. The resulting iterator will yield the exact same items as the original iterator. A close analogue from the standard library would be Iterator::inspect.

As with all iterators, the processing is lazy. The result of this method must be evaluated for the closure to be called.

fn unwrap_with<F>(self, f: F) -> UnwrapWith<Self, F> where F: FnMut(E)

Takes a closure and creates an iterator that will yield the items inside all Ok items yielded by the original iterator. All Err items will be filtered out, and the contents of each Err will be passed to the closure.

As with all iterators, the processing is lazy. The result of this method must be evaluated for the closure to be called.