Module libimagerror::trace

Traits

MapErrTrace
Helper functions for Result<T, E> types to reduce overhead in the following situations:

Functions

trace_error
Print an Error type and its cause recursively

trace_error_dbg
Print an Error type and its cause recursively with the debug!() macro

trace_error_exit
Convenience function: calls trace_error() with e and afterwards std::process::exit() with code

trace_error_maxdepth
Print an Error type and its cause recursively, but only max levels